+++ library/src/componets/books/BookCard2.jsx
@@ -4,17 +4,17 @@ import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";
 import axios from "axios";
-import { useNavigate } from "react-router-dom";
 
-const BookGallery = ({ books }) => {
+const BookGallery2 = ({ books }) => {
   const [show, setShow] = useState(false);
   const [email, setEmail] = useState(window.sessionStorage.getItem("email"));
   const [selectedBookId, setSelectedBookId] = useState("");
   const [emailError, setEmailError] = useState(false);
   const [users, setUsers] = useState([]);
+  const [selectedAction, setSelectedAction] = useState("");
   const name = window.sessionStorage.getItem("name");
 
-  const navigate = useNavigate();
+  console.log(email);
 
   useEffect(() => {
     axios
@@ -29,80 +29,67 @@ const BookGallery = ({ books }) => {
 
   const handleClose = () => {
     setShow(false);
+    setEmail("");
     setEmailError(false);
+    setSelectedAction("");
+    window.location.reload();
   };
 
   const handleShow = () => setShow(true);
 
-  const handleSubmit = async () => {
-    const emailPattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-    if (!email.trim() || !emailPattern.test(email) || !users.includes(email)) {
+  const handleSubmit = (actionType) => {
+    if (!email.trim() || !users.includes(email)) {
       setEmailError(true);
       return;
     }
 
-    try {
-      const transactionResponse = await axios.get(
-        `http://localhost:8055/api/history?email=${email}&bookId=${selectedBookId}`
-      );
-      const existingTransaction = transactionResponse.data.data;
+    const currentDate = new Date();
+    const formattedCurrentDate = currentDate.toISOString().split("T")[0];
+    let returnDate;
 
-      const existingTransactionForBook = existingTransaction.find(
-        (transaction) => transaction.bookId === selectedBookId
-      );
-
-      const currentDate = new Date();
-      const formattedCurrentDate = currentDate.toISOString().split("T")[0];
-      const returnDate = new Date(currentDate);
-      returnDate.setDate(returnDate.getDate() + 15);
-      const formattedReturnDate = returnDate.toISOString().split("T")[0];
-      const dueDate = new Date(returnDate);
-      dueDate.setDate(dueDate.getDate() + 1);
-      const formattedDueDate = dueDate.toISOString().split("T")[0];
-
-      const requestBody = {
-        email: email,
-        bookId: selectedBookId,
-        issuedDate: formattedCurrentDate,
-        dueDate: formattedDueDate,
-        returnDate: formattedReturnDate,
-        returned: false,
-      };
-
-      if (existingTransactionForBook) {
-        if (existingTransactionForBook.returned) {
-          requestBody._id = existingTransactionForBook._id;
-          await axios.put(
-            "http://localhost:8055/api/history/update",
-            requestBody
-          );
-          alert("Book transaction updated successfully!");
-          window.location.reload();
-        } else {
-          alert("You have already taken this book and not returned it yet.");
-        }
-      } else {
-        await axios.post("http://localhost:8055/api/history/add", requestBody);
-        alert("New transaction added successfully!");
-        window.location.reload();
-      }
-
-      setSelectedBookId("");
-      setEmail("");
-      setEmailError(false);
-    } catch (error) {
-      console.error("Error:", error);
+    if (actionType === "renew") {
+      const dueDate = new Date(currentDate);
+      dueDate.setDate(dueDate.getDate() + 15);
+      returnDate = dueDate.toISOString().split("T")[0];
+    } else {
+      const returnDateObj = new Date(currentDate);
+      returnDateObj.setDate(returnDateObj.getDate() + 15);
+      returnDate = returnDateObj.toISOString().split("T")[0];
     }
+
+    let requestBody = {
+      email: email,
+      bookId: selectedBookId,
+      issuedDate: formattedCurrentDate,
+      dueDate: returnDate,
+      returnDate: formattedCurrentDate,
+      returned: false,
+      renewed: false,
+    };
+
+    if (actionType === "return") {
+      requestBody.returned = true;
+      alert("One Book Returned successful!");
+    } else if (actionType === "renew") {
+      requestBody.renewed = true;
+      alert("One Book Renewed successful!");
+    }
+
+    axios
+      .put("http://localhost:8055/api/history/update", requestBody)
+      .then((response) => {
+        console.log(response.data);
+        handleClose();
+      })
+      .catch((error) => {
+        console.error("Error:", error);
+      });
   };
 
-  const handleSelectBook = (bookId) => {
-    if (email === null) {
-      alert("Please Signin to get book!");
-      navigate("/UserSignIn");
-    } else {
-      setSelectedBookId(bookId);
-      handleShow();
-    }
+  const handleSelectBook = (bookId, actionType) => {
+    setSelectedBookId(bookId);
+    setSelectedAction(actionType);
+    handleShow();
   };
 
   return (
@@ -141,9 +128,16 @@ const BookGallery = ({ books }) => {
           <Button variant="secondary" onClick={handleClose}>
             Close
           </Button>
-          <Button variant="info" onClick={handleSubmit}>
-            Submit
-          </Button>
+          {selectedAction === "return" && (
+            <Button variant="info" onClick={() => handleSubmit("return")}>
+              Return Book
+            </Button>
+          )}
+          {selectedAction === "renew" && (
+            <Button variant="info" onClick={() => handleSubmit("renew")}>
+              Renew Book
+            </Button>
+          )}
         </Modal.Footer>
       </Modal>
 
@@ -190,9 +184,20 @@ const BookGallery = ({ books }) => {
               </Typography>
               <Button
                 variant="outline-info"
-                onClick={() => handleSelectBook(book.bookId)}
+                size="sm"
+                className="ms"
+                onClick={() => handleSelectBook(book.bookId, "return")}
               >
-                Get Book
+                Return Book
+              </Button>
+
+              <Button
+                variant="outline-info"
+                size="sm"
+                className="ms"
+                onClick={() => handleSelectBook(book.bookId, "renew")}
+              >
+                Renew Book
               </Button>
             </CardContent>
           </Card>
@@ -202,4 +207,4 @@ const BookGallery = ({ books }) => {
   );
 };
 
-export default BookGallery;
+export default BookGallery2;
